@@ -3,18 +3,18 @@ part of 'package:leo_feedback/leo_feedback.dart';
 class FadeZoomBox extends StatefulWidget {
   final Widget child;
 
-  const FadeZoomBox({Key key, this.child}) : super(key: key);
+  const FadeZoomBox({Key? key, required this.child}) : super(key: key);
   @override
   _FadeZoomBoxState createState() => _FadeZoomBoxState();
 }
 
 class _FadeZoomBoxState extends State<FadeZoomBox>
     with SingleTickerProviderStateMixin {
-  AnimationController controller;
-  Animation<double> animation;
+  AnimationController? controller;
+  Animation<double>? animation;
 
   Future reverseAnimation() async {
-    await controller.reverse();
+    await controller!.reverse();
   }
 
   @override
@@ -23,12 +23,13 @@ class _FadeZoomBoxState extends State<FadeZoomBox>
     controller =
         AnimationController(vsync: this, duration: Duration(milliseconds: 100));
     animation = Tween<double>(begin: 0, end: 1)
-        .animate(CurvedAnimation(parent: controller, curve: Curves.easeIn))
+        .animate(CurvedAnimation(parent: controller!, curve: Curves.easeIn))
           ..addListener(() {
             setState(() {});
           });
-    controller.forward();
-    WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
+    controller!.forward();
+    WidgetsBinding.instance!.addPostFrameCallback((timeStamp) {
+      FeedBackHost.instance.fadeZoomBoxCreated = true;
       if (FeedBackHost.instance
               ._shouldCallAfterFadeZoomBoxWidgetCreatedFunctions.length >
           0) {
@@ -44,16 +45,17 @@ class _FadeZoomBoxState extends State<FadeZoomBox>
 
   @override
   void dispose() {
-    controller.dispose();
+    controller!.dispose();
+    FeedBackHost.instance.fadeZoomBoxCreated = false;
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
     return Transform.scale(
-      scale: animation.value,
+      scale: animation!.value,
       child: Opacity(
-        opacity: animation.value,
+        opacity: animation!.value,
         child: Center(
           child: widget.child,
         ),
